# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Talent Bridge API",
    "version": "1.0.0",
    "description": "API documentation for the Talent Bridge platform",
    "contact": {
      "name": "Talent Bridge Team",
      "email": "support@talentbridge.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
    "schemas": {
...
      "Course": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Course unique identifier",
            "example": "64f456def789abc123456789"
          },
          "title": {
            "type": "string",
            "description": "Course title",
            "example": "Web Development Fundamentals"
          },
          "instructor": {
            "type": "string",
            "description": "Instructor name",
            "example": "John Doe"
          },
          "duration": {
            "type": "string",
            "description": "Course duration",
            "example": "8 weeks"
          },
          "price": {
            "type": "number",
            "description": "Course price in PKR",
            "example": 299.99
          },
          "language": {
            "type": "string",
            "description": "Course language",
            "example": "English"
          },
          "type": {
            "type": "string",
            "enum": [
              "online",
              "offline",
              "hybrid"
            ],
            "description": "Course delivery type",
            "example": "online"
          },
          "description": {
            "type": "string",
            "description": "Course description",
            "example": "Learn the fundamentals of web development including HTML, CSS, and JavaScript"
          },
          "objectives": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Course learning objectives",
            "example": [
              "Learn HTML structure and semantics",
              "Master CSS styling and layouts",
              "Understand JavaScript fundamentals"
            ]
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Skills gained from this course",
            "example": [
              "Frontend Development",
              "Responsive Design",
              "JavaScript Programming"
            ]
          },
          "category": {
            "type": "string",
            "description": "Course category",
            "example": "Technology"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "pending_approval",
              "approved",
              "rejected",
              "archived"
            ],
            "description": "Course approval status",
            "example": "approved"
          },
          "providerId": {
            "type": "string",
            "description": "Training provider (school) ID",
            "example": "64f789abc123def456789012"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Course creation timestamp",
            "example": "2025-08-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Course last update timestamp",
            "example": "2025-08-19T14:20:00.000Z"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Always false for error responses",
            "example": false
          },
          "statusCode": {
            "type": "integer",
            "description": "HTTP error status code",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "Error message describing what went wrong",
            "example": "Validation error or operation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Detailed error messages (optional)",
            "example": [
              "Email is required",
              "Password must be at least 6 characters"
            ]
          }
        }
      },
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "description": "Current page limit",
            "example": 20
          },
          "total": {
            "type": "integer",
            "description": "Total number of items",
            "example": 150
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of pages",
            "example": 8
          },
          "hasMore": {
            "type": "boolean",
            "description": "Whether there are more results available"
          },
          "nextCursor": {
            "type": "string",
            "description": "Cursor for the next page"
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/courses": {
      "get": {
        "summary": "Get all courses with pagination and filtering",
        "description": "Retrieve a paginated list of approved courses with optional filtering by category, price range, and type",
        "tags": [
          "Courses"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination",
            "example": 1
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            },
            "description": "Number of courses per page (max 50)",
            "example": 10
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Filter by course category",
            "example": "Technology"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "online",
                "offline",
                "hybrid"
              ]
            },
            "description": "Filter by course delivery type",
            "example": "online"
          },
          {
            "in": "query",
            "name": "priceMin",
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "description": "Minimum price filter",
            "example": 100
          },
          {
            "in": "query",
            "name": "priceMax",
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "description": "Maximum price filter",
            "example": 500
          }
        ],
        "responses": {
          "200": {
            "description": "Courses retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "courses": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Course"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/PaginationInfo"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Courses retrieved successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No courses found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new course",
        "description": "Create a new course (Only available to training providers with active subscription)",
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "instructor",
                  "duration",
                  "price",
                  "language",
                  "type",
                  "description",
                  "objectives",
                  "skills",
                  "category"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Course title",
                    "example": "Web Development Fundamentals"
                  },
                  "instructor": {
                    "type": "string",
                    "description": "Instructor name",
                    "example": "John Doe"
                  },
                  "duration": {
                    "type": "string",
                    "description": "Course duration",
                    "example": "8 weeks"
                  },
                  "price": {
                    "type": "number",
                    "minimum": 0,
                    "description": "Course price in PKR",
                    "example": 299.99
                  },
                  "language": {
                    "type": "string",
                    "description": "Course language",
                    "example": "English"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "online",
                      "offline",
                      "hybrid"
                    ],
                    "description": "Course delivery type",
                    "example": "online"
                  },
                  "description": {
                    "type": "string",
                    "description": "Detailed course description",
                    "example": "Learn the fundamentals of web development including HTML, CSS, and JavaScript"
                  },
                  "objectives": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Course learning objectives",
                    "example": [
                      "Learn HTML structure and semantics",
                      "Master CSS styling and layouts",
                      "Understand JavaScript fundamentals"
                    ]
                  },
                  "skills": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Skills gained from this course",
                    "example": [
                      "Frontend Development",
                      "Responsive Design",
                      "JavaScript Programming"
                    ]
                  },
                  "category": {
                    "type": "string",
                    "description": "Course category",
                    "example": "Technology"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "course": {
                          "$ref": "#/components/schemas/Course"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Course created successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Only training providers can create courses or subscription required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Course limit exceeded for current subscription plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}